Add template support for email message
This commit adds the ability to use a custom template (ERB) for the
e-mail message with the `handler-mailer.rb` handler.

The template can be specified a few different ways.  In order of
precedence of what will be used, if specified:
  1. command-line argument (-t) to handler
  2. key in event (@event['check']['template']
  3. handler's global config (mailer.json)
  4. plugin default (backwards compatible)

By default, it will use the exact same output as it originally did,
which preserves backwards-compatibility.

The template should be an absolute path to an ERB template.

An example of a template is here:
https://gist.github.com/joshbeard/d90f2071ed69946c9b58

As a bonus, this commit also allows the timeout interval to be
customized, but defaults to the original '10'