Specify content_type
The content_type needs to be specified for sending HTML messages, which
is possible when using a custom template.

This expects the content_type to be set to 'plain' or 'html' and
defaults to 'plain' (as this handler has previously worked).

The content_type can be specified as a command-line argument, a key in
a check, or in the handler's global config.